Added Computed Numbers
So, now rather than having to put in:
  1056) One Word
  1057) Another Word

I can instead do:
  1056) One Word
  +) Another Word

And it can compute for me that the next one is 1057.
It also works supports `&)` which just uses the same number again for when the
same word has different signs for different meanings.

The risk with this setup is that the computed number will be wrong, because my
source list isn't perfect.
To help this I've built an assertion:
  1056) One Word
  +) Another Word
  =1058) A Third Word

In this case I treat it as thought it's a `+)`, but I check to make sure it has
the value I think it should.
Helps me be convenient and confident.
Good times.